fix(ui): keep entry more helper server-safe
