+++ packages/chson-ui/src/react/renderers/cheatsheet.tsx
@@ -5,7 +5,8 @@ import type { ChSONDocument } from '@chson/schema'
 import { inferColumnFormats } from '../../core/format'
 import { getLabels } from '../../core/document'
 import { getEntries, getSections } from '../../core/normalize'
-import { EntryMorePopover, hasEntryMore } from '../primitives/entry-more-popover'
+import { hasEntryMore } from '../../core/entry-more'
+import { EntryMorePopover } from '../primitives/entry-more-popover'
 import { cn } from '../utils/cn'
 import { Card, CardContent, CardHeader, CardTitle } from '../../shadcn/card'
 import { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from '../../shadcn/table'
