+++ packages/chson-ui/types/core/entry-more.d.ts
@@ -1,0 +1,3 @@
+/** True when an entry should render progressive disclosure content. */
+export declare function hasEntryMore(details?: string, url?: string): boolean;
+//# sourceMappingURL=entry-more.d.ts.map
